requirements_all.txt
numpy
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
opencv-python
scikit-learn
scikit-image
shapely
wandb
tqdm
proportion-localised
einops
efficientnet-pytorch
FrEIA
kornia
timm
imgaug
faiss-gpu